+++ piccolo/engine/cockroach.py
@@ -1,14 +1,57 @@
 from __future__ import annotations
 
 from collections.abc import Sequence
-from typing import Any, Optional
+from typing import Any, Optional, cast
 
-from piccolo.utils.lazy_loader import LazyLoader
-from piccolo.utils.warnings import Level, colored_warning
+from .postgres import Atomic, PostgresEngine, PostgresTransaction
 
-from .postgres import PostgresEngine
 
-asyncpg = LazyLoader("asyncpg", globals(), "asyncpg")
+class CockroachAtomic(Atomic):
+
+    __slots__ = ("autocommit_before_ddl",)
+
+    def __init__(
+        self,
+        engine: CockroachEngine,
+        autocommit_before_ddl: Optional[bool] = False,
+    ):
+        """
+        :param autocommit_before_ddl:
+            Defaults to ``False`` to prevent automatic DDL commits
+            in transactions (preventing rollbacks). Applies only to the current
+            transaction and is automatically reverted when the transaction
+            commits or is rolled back.
+
+            Usage::
+
+                # Defaults to ``False`` (``autocommit_before_ddl = off``)
+                transaction = engine.atomic()
+                transaction.add(Foo.create_table())
+
+                # If we want to set ``autocommit_before_ddl = on``,
+                # which is the default Cockroach session setting.
+                transaction = engine.atomic(autocommit_before_ddl=True)
+                transaction.add(Foo.create_table())
+
+        """
+        super().__init__(engine)
+        self.autocommit_before_ddl = autocommit_before_ddl
+
+    async def setup_transaction(self, transaction: PostgresTransaction):
+        if self.autocommit_before_ddl is not None:
+            transaction = cast(CockroachTransaction, transaction)
+            await transaction.autocommit_before_ddl(
+                enabled=self.autocommit_before_ddl
+            )
+
+
+class CockroachTransaction(PostgresTransaction):
+
+    async def autocommit_before_ddl(self, enabled: bool = True):
+        value = "on" if enabled else "off"
+        await self.connection.execute(
+            f"SET LOCAL autocommit_before_ddl = {value}"
+        )
 
 
 class CockroachEngine(PostgresEngine):
@@ -35,15 +78,20 @@ class CockroachEngine(PostgresEngine):
         self.engine_type = "cockroach"
         self.min_version_number = 0
 
-    async def prep_database(self):
-        try:
-            await self._run_in_new_connection(
-                "SET CLUSTER SETTING sql.defaults.experimental_alter_column_type.enabled = true;"  # noqa: E501
-            )
-        except asyncpg.exceptions.InsufficientPrivilegeError:
-            colored_warning(
-                "=> Unable to set up Cockroach DB "
-                "functionality may not behave as expected. Make sure "
-                "your database user has permission to set cluster options.",
-                level=Level.medium,
-            )
+    def atomic(
+        self,
+        autocommit_before_ddl: Optional[bool] = False,
+    ) -> CockroachAtomic:
+        return CockroachAtomic(
+            engine=self,
+            autocommit_before_ddl=autocommit_before_ddl,
+        )
+
+    def transaction(
+        self,
+        allow_nested: bool = True,
+    ) -> CockroachTransaction:
+        return CockroachTransaction(
+            engine=self,
+            allow_nested=allow_nested,
+        )